subscribe and download insights report frontend and backend

+++ app/Http/Controllers/Backend/BuilderController.php
@@ -174,6 +174,11 @@ class BuilderController extends Controller
 
     public function delete(Request $request)
     {
+        if(!isSuperAdmin()){
+            $result = ['status' => false, 'message' => 'You do not have permission to delete this builder.'];
+            return response()->json($result);
+        }
+
         $model = Builder::where('id', $request->id)->first();
         if ($model && $model->delete()) {
             $result = ['status' => true, 'message' => 'Record deleted successfully'];
